signed delta

+++ include/net/netmem_stats.h
@@ -43,7 +43,7 @@ void netmem_stats_show(struct seq_file *seq);
 void netmem_stats_show_per_site(struct seq_file *seq);
 
 void netmem_track_skb_reset(void);
-void netmem_track_skb_operation(struct sk_buff *skb, const char *operation, size_t size);
+void netmem_track_skb_operation(struct sk_buff *skb, const char *operation, long long size);
 
 /* Hardware watchpoint for truesize field */
 #ifdef CONFIG_HAVE_HW_BREAKPOINT
